feat: add output `managed_identity_principal_id` (#46)
When creating an Azure Databricks workspace, a resource group containing resources managed by the Databricks platform. One of these resources is a managed identity named `dbmanagedidentity`, which is assigned to the virtual machines that the Databricks clusters run on.

By outputting the object ID of this managed identity, it can be assigned additional roles, e.g. "Key Vault Secrets User" to read Key Vault secrets.